+++ macro/src/main/scala/com/github/phisgr/gatling/ForToMatch.scala
@@ -18,8 +18,11 @@ object ForToMatch {
     tree match {
       case Apply(TypeApply(Select(v, TermName("flatMap")), _), List(Function(List(ValDef(_, name, _, _)), body))) =>
         matchValidation(v, name, toMatch(c)(body))
-      case Apply(TypeApply(Select(v, TermName("map")), _), List(Function(List(ValDef(_, name, _, _)), body))) =>
-        matchValidation(v, name, q"io.gatling.commons.validation.Success { $body }")
+      case Apply(TypeApply(Select(v, TermName("map")), _), List(Function(List(ValDef(_, name, tpt, _)), body))) =>
+        body match {
+          case Literal(Constant(())) if tpt.tpe =:= definitions.UnitTpe => v
+          case _ => matchValidation(v, name, q"io.gatling.commons.validation.Success { $body }")
+        }
       case tree =>
         c.warning(c.enclosingPosition, s"Early stopping with ${tree.getClass}:\n$tree")
         tree