+++ src/main/scala/com/github/phisgr/gatling/grpc/action/GrpcCallActionBuilder.scala
@@ -1,7 +1,7 @@
 package com.github.phisgr.gatling.grpc.action
 
-import com.github.phisgr.gatling.grpc.check.{CheckMixin, GrpcCheck}
-import com.github.phisgr.gatling.grpc.request.{CallAttributes, CallAttributesMixin}
+import com.github.phisgr.gatling.grpc.check.GrpcCheck
+import com.github.phisgr.gatling.grpc.request.{CallAttributes, CallDefinition}
 import io.gatling.core.action.Action
 import io.gatling.core.action.builder.ActionBuilder
 import io.gatling.core.session.Expression
@@ -10,14 +10,13 @@ import io.grpc.MethodDescriptor
 
 case class GrpcCallActionBuilder[Req, Res](
   requestName: Expression[String],
-  method: MethodDescriptor[Req, Res],
+  private[gatling] override val method: MethodDescriptor[Req, Res],
   payload: Expression[Req],
-  private[gatling] val callAttributes: CallAttributes = CallAttributes(),
-  checks: List[GrpcCheck[Res]] = Nil,
+  private[gatling] override val callAttributes: CallAttributes = CallAttributes(),
+  private[gatling] override val checks: List[GrpcCheck[Res]] = Nil,
   isSilent: Boolean = false
 ) extends ActionBuilder
-  with CallAttributesMixin[GrpcCallActionBuilder[Req, Res]]
-  with CheckMixin[GrpcCallActionBuilder[Req, Res], GrpcCheck, Res] {
+  with CallDefinition[GrpcCallActionBuilder[Req, Res], GrpcCheck, Req, Res] {
 
   override def build(ctx: ScenarioContext, next: Action): Action = new GrpcCallAction(this, ctx, next)