+++ src/main/scala/com/github/phisgr/gatling/grpc/action/serverStream.scala
@@ -2,33 +2,31 @@ package com.github.phisgr.gatling.grpc.action
 
 import com.github.phisgr.gatling.forToMatch
 import com.github.phisgr.gatling.grpc.check.GrpcResponse.GrpcStreamEnd
-import com.github.phisgr.gatling.grpc.check.{CheckMixin, GrpcCheck, StatusExtract, StreamCheck}
-import com.github.phisgr.gatling.grpc.request.{Call, CallAttributes, CallAttributesMixin}
+import com.github.phisgr.gatling.grpc.check.StreamCheck
+import com.github.phisgr.gatling.grpc.request.CallAttributes
 import com.github.phisgr.gatling.grpc.stream.StreamCall.ensureNoStream
 import com.github.phisgr.gatling.grpc.stream.{ServerStreamCall, SessionCombiner, TimestampExtractor}
 import io.gatling.commons.util.Clock
 import io.gatling.commons.validation.Validation
+import io.gatling.core.action.Action
 import io.gatling.core.action.builder.ActionBuilder
-import io.gatling.core.action.{Action, RequestAction}
 import io.gatling.core.session.{Expression, Session}
 import io.gatling.core.stats.StatsEngine
 import io.gatling.core.structure.ScenarioContext
-import io.gatling.core.util.NameGen
 import io.grpc.MethodDescriptor
 
 case class ServerStreamStartActionBuilder[Req, Res](
   private[gatling] val requestName: Expression[String],
   private[gatling] val streamName: String,
-  private[gatling] val method: MethodDescriptor[Req, Res],
+  private[gatling] override val method: MethodDescriptor[Req, Res],
   private[gatling] val req: Expression[Req],
-  private[gatling] val _timestampExtractor: TimestampExtractor[Res] = TimestampExtractor.Ignore,
+  private[gatling] override val _timestampExtractor: TimestampExtractor[Res] = TimestampExtractor.Ignore,
   private[gatling] val _sessionCombiner: SessionCombiner = SessionCombiner.NoOp,
-  private[gatling] val callAttributes: CallAttributes = CallAttributes(),
-  private[gatling] val checks: List[StreamCheck[Res]] = Nil,
-  private[gatling] val endChecks: List[StreamCheck[GrpcStreamEnd]] = Nil,
+  private[gatling] override val callAttributes: CallAttributes = CallAttributes(),
+  private[gatling] override val checks: List[StreamCheck[Res]] = Nil,
+  private[gatling] override val endChecks: List[StreamCheck[GrpcStreamEnd]] = Nil,
 ) extends ActionBuilder
-  with CallAttributesMixin[ServerStreamStartActionBuilder[Req, Res]]
-  with CheckMixin[ServerStreamStartActionBuilder[Req, Res], StreamCheck, Res] {
+  with StreamStartBuilder[ServerStreamStartActionBuilder[Req, Res], StreamCheck, Req, Res] {
 
   override def build(ctx: ScenarioContext, next: Action): Action =
     new ServerStreamStartAction(this, ctx, next)
@@ -51,11 +49,9 @@ class ServerStreamStartAction[Req, Res](
   builder: ServerStreamStartActionBuilder[Req, Res],
   ctx: ScenarioContext,
   override val next: Action
-) extends Call(ctx, builder.callAttributes) with RequestAction with NameGen {
+) extends StreamStartAction[Req, Res](ctx, builder) {
 
-  private[this] val endChecks = if (builder.endChecks.exists(_.scope == GrpcCheck.Status)) builder.endChecks else {
-    StatusExtract.DefaultStreamCheck :: builder.endChecks
-  }
+  override protected def callClass: Class[_] = classOf[ServerStreamCall[_, _]]
 
   override def requestName: Expression[String] = builder.requestName
   override def sendRequest(requestName: String, session: Session): Validation[Unit] = forToMatch {
@@ -67,11 +63,11 @@ class ServerStreamStartAction[Req, Res](
       resolvedPayload <- builder.req(session)
       callOptions <- callOptions(session)
     } yield {
-      val channel = component.getChannel(session)
+      val call = newCall(session, callOptions)
       next ! session.set(streamName, new ServerStreamCall(
         requestName = requestName,
         streamName = streamName,
-        channel.newCall(builder.method, callOptions),
+        call,
         headers,
         resolvedPayload,
         ctx,
@@ -79,7 +75,8 @@ class ServerStreamStartAction[Req, Res](
         builder._sessionCombiner,
         session,
         builder.checks,
-        endChecks
+        endChecks,
+        ignoreMessage = ignoreMessage
       ))
     }
   }