+++ src/main/scala/com/github/phisgr/gatling/grpc/request/Call.scala
@@ -1,12 +1,18 @@
 package com.github.phisgr.gatling.grpc.request
 
-import com.github.phisgr.gatling.grpc.protocol.GrpcProtocol
+import com.github.phisgr.gatling.grpc.protocol.{ByteArrayMarshaller, EmptyMarshaller, GrpcProtocol}
 import io.gatling.commons.validation.{SuccessWrapper, Validation}
+import io.gatling.core.action.RequestAction
 import io.gatling.core.session.{Expression, Session}
 import io.gatling.core.structure.ScenarioContext
-import io.grpc.{CallOptions, Metadata}
+import io.gatling.core.util.NameGen
+import io.grpc.{CallOptions, ClientCall, Metadata, MethodDescriptor}
 
-abstract class Call(ctx: ScenarioContext, callAttributes: CallAttributes) {
+abstract class Call[Req, Res](
+  ctx: ScenarioContext,
+  callAttributes: CallAttributes,
+  method: MethodDescriptor[Req, Res]
+) extends RequestAction with NameGen {
 
   private[this] val headerPairs = callAttributes.reversedHeaders.reverse.toArray
   protected def resolveHeaders(session: Session): Validation[Metadata] = {
@@ -23,8 +29,34 @@ abstract class Call(ctx: ScenarioContext, callAttributes: CallAttributes) {
 
   protected val callOptions: Expression[CallOptions] = callAttributes.callOptions
 
-  protected val component: GrpcProtocol.GrpcComponent = {
+  private[this] val component: GrpcProtocol.GrpcComponent = {
     val protocolKey = callAttributes.protocolOverride.fold(GrpcProtocol.GrpcProtocolKey)(_.overridingKey)
     ctx.protocolComponentsRegistry.components(protocolKey)
+  }
+
+  protected def needParsed: Boolean
+  protected def mayNeedDelayedParsing: Boolean
+
+  protected val lazyParseMethod: MethodDescriptor[Req, Any] = {
+    val notParsed = component.lazyParsing && !needParsed
+    val withResponseMarshaller = if (notParsed) {
+      val responseMarshaller = if (mayNeedDelayedParsing) {
+        ByteArrayMarshaller
+      } else {
+        EmptyMarshaller
+      }
+      method.toBuilder(
+        method.getRequestMarshaller,
+        responseMarshaller
+      ).build()
+    } else {
+      method
+    }
+
+    withResponseMarshaller.asInstanceOf[MethodDescriptor[Req, Any]]
+  }
+
+  protected def newCall(session: Session, callOptions: CallOptions): ClientCall[Req, Any] = {
+    component.getChannel(session).newCall(lazyParseMethod, callOptions)
   }
 }