+++ src/test/scala/com/github/phisgr/example/StreamingExample.scala
@@ -9,7 +9,9 @@ import com.github.phisgr.gatling.grpc.Predef._
 import com.github.phisgr.gatling.grpc.stream.{SessionCombiner, TimestampExtractor}
 import com.github.phisgr.gatling.pb._
 import com.google.protobuf.empty.Empty
+import io.gatling.commons.validation.{Failure, Validation}
 import io.gatling.core.Predef._
+import io.gatling.core.check.Matcher
 import io.gatling.core.session.Expression
 import io.grpc.{CallOptions, Status}
 
@@ -27,7 +29,7 @@ class StreamingExample extends Simulation {
     .serverStream(streamName = "listener")
   val chatCall = grpc("Chat")
     .bidiStream(streamName = "chatter")
-  val complete = chatCall.copy("Complete").complete
+  val complete = chatCall.copy(requestName = "Complete").complete
 
   val grpcConf = grpc(managedChannelBuilder(target = "localhost:8080").usePlaintext())
     .shareChannel
@@ -53,17 +55,7 @@ class StreamingExample extends Simulation {
           session
         }
     }
-    .exec(
-      grpc("Listen")
-        .serverStream(streamName = "listener")
-        .cancelStream
-    )
-    .exec(
-      listenCall
-        .copy(requestName = "Cannot build")
-        .start(ChatServiceGrpc.METHOD_LISTEN)(Empty.defaultInstance)
-        .header(TokenHeaderKey)($("whatever"))
-    )
+    .exec(listenCall.cancelStream)
 
   val chatter = scenario("Chatter")
     .exec(_.set("username", UUID.randomUUID().toString))
@@ -106,16 +98,33 @@ class StreamingExample extends Simulation {
         )
     }
     .exec(complete)
+    .exec(chatCall.copy(requestName = "Send after complete").send(ChatMessage.defaultInstance))
     .exec(complete)
 
+
+  val endsWithHi: Matcher[String] = new Matcher[String] {
+    override protected def doMatch(actual: Option[String]): Validation[Option[String]] =
+      if (actual.exists(_.endsWith("hi"))) actual else {
+        val actualEnding = actual.map(_.split(' ').last)
+        Failure(s"ends with $actualEnding")
+      }
+
+    override def name: String = "endsWith(hi)"
+  }
   val failure = scenario("Failures")
     .feed(List(true, false).map(b => Map("earlyStop" -> b)).iterator)
+    .exec(
+      listenCall
+        .copy(requestName = "Cannot build")
+        .start(ChatServiceGrpc.METHOD_LISTEN)(Empty.defaultInstance)
+        .header(TokenHeaderKey)($("whatever"))
+    )
     .exec(chatCall.copy("Fail")
       .connect(ChatServiceGrpc.METHOD_CHAT)
       .timestampExtractor { (_, message, _) =>
         if (ThreadLocalRandom.current().nextBoolean()) message.time - 100 else throw new IllegalStateException()
       }
-      .extract(_.data.some)(_ transform (_.endsWith("hi")) is true)
+      .extract(_.data.some)(_ validate endsWithHi)
       .endCheck(statusCode is Status.Code.CANCELLED)
     )
     .pause(2.seconds)